fixed compiler warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@53998 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 2782626f011519eb3302a826bf34103b3771c37b

+++ Rec/Rich/RichHPDImageAnalysis/src/RichHPDImageSummary.cpp
@@ -215,9 +215,9 @@ void RichHPDImageSummary::summaryINFO( const unsigned int ID,
                                        const TH2D* hist ) const
 {
 
-  const unsigned int nPix = hist->Integral();
+  const unsigned int nPix = (unsigned int) (hist->Integral());
   if ( nPix < m_minOccupancy ) return ;
-
+  
   HPDBoundaryFcn FCN( hist , m_cutFraction );
 
   const int boundarySize = FCN.findBoundary() ;
